Category api tests

+++ src/Services/Category/ContestPark.Category.API/Controllers/SubCategoryController.cs
@@ -2,17 +2,15 @@
 using ContestPark.Category.API.Infrastructure.Repositories.Category;
 using ContestPark.Category.API.Infrastructure.Repositories.FollowSubCategory;
 using ContestPark.Category.API.Infrastructure.Repositories.OpenCategory;
-using ContestPark.Category.API.Infrastructure.Repositories.Search;
 using ContestPark.Category.API.Model;
 using ContestPark.Category.API.Resources;
 using ContestPark.Core.CosmosDb.Models;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using System;
+using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
-using static ContestPark.Category.API.Infrastructure.Repositories.Search.SearchRepository;
 
 namespace ContestPark.Category.API.Controllers
 {
@@ -23,7 +21,6 @@ namespace ContestPark.Category.API.Controllers
         private readonly IFollowSubCategoryRepository _followSubCategoryRepository;
         private readonly IOpenCategoryRepository _openCategoryRepository;
         private readonly ICategoryRepository _categoryRepository;
-        private readonly ISearchRepository _searchRepository;
 
         #endregion Private Variables
 
@@ -32,13 +29,11 @@ namespace ContestPark.Category.API.Controllers
         public SubCategoryController(IFollowSubCategoryRepository followSubCategoryRepository,
                                      IOpenCategoryRepository openCategoryRepository,
                                      ICategoryRepository categoryRepository,
-                                     ISearchRepository searchRepository,
                                      ILogger<SubCategoryController> logger) : base(logger)
         {
             _followSubCategoryRepository = followSubCategoryRepository ?? throw new ArgumentNullException(nameof(followSubCategoryRepository));
             _openCategoryRepository = openCategoryRepository ?? throw new ArgumentNullException(nameof(openCategoryRepository));
             _categoryRepository = categoryRepository ?? throw new ArgumentNullException(nameof(categoryRepository));
-            _searchRepository = searchRepository ?? throw new ArgumentNullException(nameof(searchRepository));
         }
 
         #endregion Constructor
@@ -49,15 +44,20 @@ namespace ContestPark.Category.API.Controllers
         /// Kategorileri listeleme
         /// </summary>
         /// <returns>Tüm kategorileri döndürür.</returns>
+        /// <param name="pagingModel">Hem kategorileri hemde alt kategorileri sayfalar</param>
         [HttpGet]
         [ProducesResponseType(typeof(ServiceModel<CategoryModel>), (int)HttpStatusCode.OK)]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
         public IActionResult Get([FromQuery]PagingModel pagingModel)
         {
-            ServiceModel<CategoryModel> catgories = _categoryRepository.GetCategories(UserId, CurrentUserLanguage, pagingModel);
-            if (catgories == null)
+            ServiceModel<CategoryModel> catgories = _categoryRepository.GetCategories(UserId,
+                                                                                      CurrentUserLanguage,
+                                                                                      pagingModel);
+
+            if (catgories == null || catgories.Items == null || catgories.Items.Count() == 0)
             {
-                Logger.LogCritical($"{nameof(catgories)} list returned empty.", catgories);
+                Logger.LogCritical($"{nameof(catgories)} list returned empty.");
+
                 return NotFound();
             }
 
@@ -69,14 +69,17 @@ namespace ContestPark.Category.API.Controllers
         /// </summary>
         /// <returns>Tüm kategorileri döndürür.</returns>
         [HttpGet("Followed")]
-        [ProducesResponseType(typeof(ServiceModel<CategoryModel>), (int)HttpStatusCode.OK)]
+        [ProducesResponseType(typeof(ServiceModel<SubCategoryModel>), (int)HttpStatusCode.OK)]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
         public IActionResult FollowedSubCategories([FromQuery]PagingModel pagingModel)
         {
-            ServiceModel<SubCategoryModel> followedSubCategories = _categoryRepository.GetFollowedSubCategories(UserId, CurrentUserLanguage, pagingModel);
-            if (followedSubCategories == null)
+            ServiceModel<SubCategoryModel> followedSubCategories = _categoryRepository.GetFollowedSubCategories(UserId,
+                                                                                                                CurrentUserLanguage,
+                                                                                                                pagingModel);
+
+            if (followedSubCategories == null || followedSubCategories.Items == null || followedSubCategories.Items.Count() == 0)
             {
-                Logger.LogWarning($"{nameof(followedSubCategories)} list returned empty.", followedSubCategories);
+                Logger.LogInformation($"{nameof(followedSubCategories)} list returned empty.");
 
                 return NotFound();
             }
@@ -85,59 +88,13 @@ namespace ContestPark.Category.API.Controllers
         }
 
         /// <summary>
-        /// Takip ettiğin alt kategoriler
-        /// </summary>
-        /// <returns>Tüm kategorileri döndürür.</returns>
-        [HttpGet("Followed/search")]
-        [ProducesResponseType(typeof(ServiceModel<SearchModel>), (int)HttpStatusCode.OK)]
-        [ProducesResponseType((int)HttpStatusCode.BadRequest)]
-        public async Task<IActionResult> FollowedSubCategories([FromQuery(Name = "q")]string searchText, [FromQuery]PagingModel pagingModel)
-        {
-            ServiceModel<SearchModel> followedSearchSubCategories = await _searchRepository.SearchFollowedSubCategoriesAsync(searchText, UserId, CurrentUserLanguage, pagingModel);
-            if (followedSearchSubCategories == null)
-            {
-                Logger.LogWarning($"{nameof(followedSearchSubCategories)} list returned empty.", followedSearchSubCategories);
-
-                return NotFound();
-            }
-
-            return Ok(followedSearchSubCategories);
-        }
-
-        /// <summary>
-        /// Kategori arama
-        /// </summary>
-        /// <returns>Tüm kategorileri döndürür.</returns>
-        [HttpGet("search/{categoryId?}")]
-        [AllowAnonymous]
-        [ProducesResponseType(typeof(ServiceModel<SearchModel>), (int)HttpStatusCode.OK)]
-        [ProducesResponseType((int)HttpStatusCode.BadRequest)]
-        public async Task<IActionResult> SearchAsync(string categoryId, [FromQuery(Name = "q")]string searchText, [FromQuery]PagingModel pagingModel)
-        {
-            ServiceModel<SearchModel> searchCategories = await _searchRepository.DynamicSearchAsync(searchText,
-                                                                                                    CurrentUserLanguage,
-                                                                                                    pagingModel,
-                                                                                                    SearchFilters.CategoryId,
-                                                                                                    categoryId);
-            if (searchCategories == null)
-            {
-                Logger.LogWarning($"{nameof(searchCategories)} list returned empty.", searchCategories);
-
-                return NotFound();
-            }
-
-            return Ok(searchCategories);
-        }
-
-        /// <summary>
         /// Alt kategori takip et
         /// </summary>
         /// <param name="subCategoryId">Alt kategori Id</param>
-        [HttpPost]
-        [Route("{subCategoryId}/Follow")]
+        [HttpPost("{subCategoryId}/Follow")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
-        public async Task<IActionResult> Post(string subCategoryId)
+        public async Task<IActionResult> Post([FromRoute]string subCategoryId)
         {
             if (string.IsNullOrEmpty(subCategoryId) || string.IsNullOrEmpty(UserId))
                 return BadRequest();
@@ -145,11 +102,15 @@ namespace ContestPark.Category.API.Controllers
             if (_followSubCategoryRepository.IsSubCategoryFollowed(UserId, subCategoryId))// Kategoriyi daha önceden takip etmişmi
                 return BadRequest(CategoryResource.YouAreAlreadyFollowingThisCategory);
 
-            if (!(_categoryRepository.IsSubCategoryFree(subCategoryId) ||// kategori ücretsiz değil ise
-                _openCategoryRepository.IsSubCategoryOpen(UserId, subCategoryId)))// veya kategorinin kilidi açık değilse
+            if (!(
+                _categoryRepository.IsSubCategoryFree(subCategoryId) ||// kategori ücretsiz değil ise
+                _openCategoryRepository.IsSubCategoryOpen(UserId, subCategoryId)// veya kategorinin kilidi açık değilse
+                ))
+            {
                 return BadRequest(CategoryResource.ToBeAbleToFollowThisCategoryYouNeedToUnlockIt);
+            }
 
-            bool isSuccess = await _followSubCategoryRepository.AddAsync(new Infrastructure.Documents.FollowSubCategory
+            bool isSuccess = await _followSubCategoryRepository.AddAsync(new FollowSubCategory
             {
                 SubCategoryId = subCategoryId,
                 UserId = UserId
@@ -179,10 +140,13 @@ namespace ContestPark.Category.API.Controllers
         [Route("{subCategoryId}/UnFollow")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
-        public async Task<IActionResult> Delete(string subCategoryId)
+        public async Task<IActionResult> Delete([FromRoute]string subCategoryId)
         {
             if (string.IsNullOrEmpty(subCategoryId) || string.IsNullOrEmpty(UserId))
                 return BadRequest();
+
+            if (!_followSubCategoryRepository.IsSubCategoryFollowed(UserId, subCategoryId))
+                return BadRequest(CategoryResource.YouMustFollowThisCategoryToDeactivateTheCategory);
 
             bool isSuccess = await _followSubCategoryRepository.DeleteAsync(UserId, subCategoryId);
             if (!isSuccess)
@@ -210,7 +174,7 @@ namespace ContestPark.Category.API.Controllers
         [Route("{subCategoryId}/FollowStatus")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
-        public IActionResult Get(string subCategoryId)
+        public IActionResult Get([FromRoute]string subCategoryId)
         {
             if (string.IsNullOrEmpty(subCategoryId) || string.IsNullOrEmpty(UserId))
                 return BadRequest();
@@ -229,7 +193,7 @@ namespace ContestPark.Category.API.Controllers
         [Route("{subCategoryId}")]
         [ProducesResponseType(typeof(SubCategoryDetailModel), (int)HttpStatusCode.OK)]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
-        public IActionResult GetDetail(string subCategoryId)
+        public IActionResult GetDetail([FromRoute]string subCategoryId)
         {
             if (string.IsNullOrEmpty(subCategoryId) || string.IsNullOrEmpty(UserId))
                 return BadRequest();
@@ -257,18 +221,23 @@ namespace ContestPark.Category.API.Controllers
         /// Alt kategori kilit açma
         /// </summary>
         /// <param name="subCategoryId">Alt kategori Id</param>
-        [HttpPost]
-        [Route("{subCategoryId}/unlock")]
+        [HttpPost("{subCategoryId}/unlock")]
         [ProducesResponseType((int)HttpStatusCode.OK)]
         [ProducesResponseType((int)HttpStatusCode.BadRequest)]
-        public async Task<IActionResult> UnLockSubCategory(string subCategoryId)
+        public async Task<IActionResult> UnLockSubCategory([FromRoute]string subCategoryId)
         {
             if (string.IsNullOrEmpty(subCategoryId) || string.IsNullOrEmpty(UserId))
                 return BadRequest();
 
+            bool isSubCategoryFree = _categoryRepository.IsSubCategoryFree(subCategoryId);
+            if (isSubCategoryFree)
+                return BadRequest(CategoryResource.YouCanNotUnlockTheFreeCategory);
+
             bool isSubCategoryOpen = _openCategoryRepository.IsSubCategoryOpen(UserId, subCategoryId);
             if (isSubCategoryOpen)
                 return BadRequest(CategoryResource.ThisCategoryIsAlreadyUnlocked);
+
+            // TODO: check altın miktarı yeterli mi
 
             bool isSuccess = await _openCategoryRepository.AddAsync(new OpenSubCategory
             {
